return kernel and syscall contexts to free list of originating cpu
The prior scheme for recycling kernel and syscall contexts was naively
returning retired contexts to the free list of the cpu where the final release
occurred. This was leading to an imbalance of cached contexts amongst cpus and
an excess of contexts being allocated to compensate. To remedy this, use the
lock-free queue to return freed contexts to the cpu from which they
originated. This creates an upper bound to the amount of context data that may
be cached per processor (adjustable via constants in config.h). Contexts that
cannot be returned to full free queues are simply deallocated. Note that these
queues have multiple producers but only a single consumer (the owning cpu), so
the more optimal dequeue_single() is used when removing contexts from the
queue.